Players now store all collected weapons/armors
Added armors[] and weapons[] arrays to player.js, and modified how
weapon gathering is handled.  The arrays hold the type property of the
weapons, and multiple of that type may be carried.  Also added the
saving procedure for those arrays into game.js and storage.js.

Next is being able to select equipment from a list of carried.